# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Music API 🎵",
    "description": "API quản lý bài hát, playlist và nghệ sĩ.",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/artists/": {
      "post": {
        "tags": [
          "Artist"
        ],
        "description": "",
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "stageName": {
                    "example": "any"
                  },
                  "bio": {
                    "example": "any"
                  },
                  "youtubeUrl": {
                    "example": "any"
                  },
                  "facebookUrl": {
                    "example": "any"
                  },
                  "instagramUrl": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Artist"
        ],
        "description": "",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "size",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/artists/me": {
      "get": {
        "tags": [
          "Artist"
        ],
        "description": "",
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "put": {
        "tags": [
          "Artist"
        ],
        "description": "",
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "stageName": {
                    "example": "any"
                  },
                  "bio": {
                    "example": "any"
                  },
                  "youtubeUrl": {
                    "example": "any"
                  },
                  "facebookUrl": {
                    "example": "any"
                  },
                  "instagramUrl": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/artists/myFollowers": {
      "get": {
        "tags": [
          "Artist"
        ],
        "description": "",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "size",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/artists/follows": {
      "get": {
        "tags": [
          "Artist"
        ],
        "description": "",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "size",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/artists/{id}": {
      "get": {
        "tags": [
          "Artist"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/artists/follow/{id}": {
      "post": {
        "tags": [
          "Artist"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/artists/unfollow/{id}": {
      "delete": {
        "tags": [
          "Artist"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/songs/artist/{artistId}": {
      "get": {
        "tags": [
          "Song"
        ],
        "description": "",
        "parameters": [
          {
            "name": "artistId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "size",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/nfts/set-fee": {
      "post": {
        "tags": [
          "NFT"
        ],
        "description": "",
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "newFee": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/nfts/set-wallet": {
      "post": {
        "tags": [
          "NFT"
        ],
        "description": "",
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "newWallet": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/nfts/create-ticket": {
      "post": {
        "tags": [
          "NFT"
        ],
        "description": "",
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "baseUrl": {
                    "example": "any"
                  },
                  "contractAddress": {
                    "example": "any"
                  },
                  "date": {
                    "example": "any"
                  },
                  "saleDeadline": {
                    "example": "any"
                  },
                  "price": {
                    "example": "any"
                  },
                  "location": {
                    "example": "any"
                  },
                  "title": {
                    "example": "any"
                  },
                  "maxSupply": {
                    "example": "any"
                  },
                  "coverImage": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/nfts/log-purchase": {
      "post": {
        "tags": [
          "NFT"
        ],
        "description": "",
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "eventId": {
                    "example": "any"
                  },
                  "tokenId": {
                    "example": "any"
                  },
                  "txHash": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/nfts/update-status/{id}": {
      "post": {
        "tags": [
          "NFT"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/nfts": {
      "get": {
        "tags": [
          "NFT"
        ],
        "description": "",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/nfts/my-ticket": {
      "get": {
        "tags": [
          "NFT"
        ],
        "description": "",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/nfts/artist-ticket": {
      "get": {
        "tags": [
          "NFT"
        ],
        "description": "",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/nfts/resell": {
      "post": {
        "tags": [
          "NFT"
        ],
        "description": "",
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userTicketId": {
                    "example": "any"
                  },
                  "price": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "NFT"
        ],
        "description": "",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/nfts/resell/buy": {
      "post": {
        "tags": [
          "NFT"
        ],
        "description": "",
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "resellTicketId": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/nfts/resell/{resellId}": {
      "patch": {
        "tags": [
          "NFT"
        ],
        "description": "",
        "parameters": [
          {
            "name": "resellId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}